+++ src/main/java/net/spell_engine/client/SpellEngineClient.java
@@ -10,6 +10,9 @@ import net.minecraft.item.BowItem;
 import net.minecraft.item.CrossbowItem;
 import net.minecraft.registry.Registries;
 import net.spell_engine.SpellEngineMod;
+import net.spell_engine.api.effect.CustomParticleStatusEffect;
+import net.spell_engine.api.effect.SpellEngineEffects;
+import net.spell_engine.api.render.StunParticleSpawner;
 import net.spell_engine.client.compatibility.CompatFeatures;
 import net.spell_engine.client.render.BeamRenderer;
 import net.spell_engine.client.render.ModelPredicateHelper;
@@ -60,4 +63,10 @@ public class SpellEngineClient {
             }
         }
     }
+
+    private static void registerEffectParticles() {
+        CustomParticleStatusEffect.register(
+                SpellEngineEffects.STUN.effect,
+                new StunParticleSpawner()
+        );}
 }